day09/lab/pyproject.toml
[project]
name = "lab"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.7",
    "google-cloud-aiplatform>=1.147.0",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "tqdm>=4.67.3",
]
